% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetHistoricTradeData}
\alias{GetHistoricTradeData}
\title{This function will return the 2000 most recent trades which are delayed by
approximately 15 minutes.}
\usage{
GetHistoricTradeData(params)
}
\arguments{
\item{params}{Any parameter accepted by this web service call -- see
\href{http://bitcoincharts.com/about/markets-api/}{here}.}
}
\description{
The symbols that are available can be found \href{http://bitcoincharts.com/markets/}{here}.
}
\details{
Note that calling this function with invalid parameters will result in an
empty data frame.
}
\examples{
params <- list (symbol="btceUSD")
 historicData <- GetHistoricTradeData (params)
}

