% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetMarketData}
\alias{GetMarketData}
\title{This function will return an array with elements for each market.}
\usage{
GetMarketData(params)
}
\arguments{
\item{params}{Any parameter accepted by this web service call -- see \href{http://bitcoincharts.com/about/markets-api/}{here}}
}
\description{
General market data can be accessed \href{http://api.bitcoincharts.com/v1/markets.json}{here}.
}
\examples{
params <- list (currency="USD")
 tryCatch(
     usd <- GetMarketData (params),
     error =
         function (e) {
             print (
                 paste (
                     "An exception was thrown -- details follow: ",
                     e,
                     sep=""
                 )
             )
         }
     )
}

