% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetWeightedPrices}
\alias{GetWeightedPrices}
\title{This function returns the weighted prices.}
\usage{
GetWeightedPrices()
}
\value{
Weighted prices are calculated for the last 24 hours, 7 days and 30
 days; if there are no trades during an interval, such as no trade within 24
 hours, then no value will be returned.
}
\description{
\href{http://www.bitcoincharts.com}{Bitcoincharts} offers weighted prices
for several currencies that can be used, for example, to price goods and
services in Bitcoins -- this will yield much lower fluctuations than using
a single market's latest price.
}
\examples{
tryCatch(
     weightedPrices <- GetWeightedPrices (),
     Throwable =
         function (e) {
             print (
                 paste (
                     "An exception was thrown -- details follow.",
                     e$getMessage(),
                     sep=""
                 )
             )
         }
     )
}

