\name{GetHistoricTradeData}
\alias{GetHistoricTradeData}
\title{GetHistoricTradeData}
\description{
    Function returns a data frame of the most recent trades.
}
\arguments{
    See http://bitcoincharts.com/about/markets-api/ for a list of available
    parameters.
}
\value{
    A data frame containing the historic trade data.
}